% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjcolor.R
\name{adjColor}
\alias{adjColor}
\title{adjColor}
\usage{
adjColor(col, alpha = 1, bg = "#FFFFFF")
}
\arguments{
\item{col}{adding colour}

\item{alpha}{the opacity of alpha, in the range [0,1]}

\item{bg}{the background colour used as a reference. Default is white,}
}
\description{
Adds two colours together assuming a background colour bg. Acts in a similar way to \code{\link{adjustcolor}} with the exception that alpha is not used. Alpha components are lost when sending coloured text or tables to word through markdown or ReporterRs.
}
