% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankPlot.R
\name{blankPlot}
\alias{blankPlot}
\title{blank Plots}
\usage{
blankPlot(xlim = c(0, 1), ylim = c(0, 1), xaxs = "i", yaxs = "i")
}
\arguments{
\item{xlim}{The x axis range}

\item{ylim}{The y axis range}

\item{xaxs}{Either "i" or "r". i fits the range exactly whilst r add a little white space to either end. Default is i here which is the reverse of plot().}

\item{yaxs}{Either "i" or "r". i fits the range exactly whilst r add a little white space to either end. Default is i here which is the reverse of plot().}
}
\description{
This is a convinience wrapper function for creating a blank plot. You only need to suppy limits to initialise the plot. Then add to the plot using axis, points, lines, mtext, legend etc.
}
