% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_boxDensityPlot.R
\name{boxDensityPlot}
\alias{boxDensityPlot}
\title{boxDensityPlot}
\usage{
boxDensityPlot(dta, diff = 0.1, rnd = 0.01, xaxs = "i", yaxs = "i",
  ylab = "Density", xlab = "test", col = 2, border = 1, add = FALSE,
  xlim = NULL, ylim = NULL, ...)
}
\arguments{
\item{dta}{The data}

\item{diff}{The width of each item to be plotted in}

\item{rnd}{a number to round all values to}

\item{xaxs, yaxs, xlim, ylim, ylab}{Standard plot options}

\item{col}{the fill colour}

\item{border}{the border colour}

\item{add}{TRUE/FALSE add to an existing plot or not}

\item{...}{other values passed to the polygon function}
}
\description{
This is an alternative plot to hist or plot density. Instead of binning or smoothing each item is drawn around its position +- diff. These are then stacked vertically. A rounding option is built in to give a cleaner building block look.
}

