% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_queryList_referenceClass.R
\docType{class}
\name{dataQueries-class}
\alias{dataQueries-class}
\title{dataQueries reference class}
\description{
A reference class for dataqueries to prevent the need to return the data back to global.
}
\section{Fields}{

\describe{
\item{\code{q}}{data.frame containing these columns id, identifier, queryRun, patid, CRF, repeatLine, firstQuery, STATScomments,comments, resolved, STATSresolved}

\item{\code{queryRun}}{A number denoting the query run. Useful for seeing what queries are still active.}
}}

