% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandColours.R
\name{expandColours}
\alias{expandColours}
\title{expand Colours}
\usage{
expandColours(cols, count)
}
\arguments{
\item{cols}{a vector of hex colours.}

\item{count}{the number of colours to return.}
}
\description{
A function for linearly interpolating a vector of colour
}
