% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenClinica_importOpenClinica.R
\name{importOpenClinica}
\alias{importOpenClinica}
\alias{importOpenClinica_batch}
\title{import Open Clinica data}
\usage{
importOpenClinica(txt, path, dates = NULL)

importOpenClinica_batch(path, trialStub, dbName, newFolder, delStub = "")
}
\arguments{
\item{txt}{Filename}

\item{path}{Filepath}

\item{dates}{A vector of column names to convert to dates}

\item{trialStub}{A string to filter unwanted files from the target folder}

\item{dbName}{The file name for the .Rdata file will all the data contained within it}

\item{delStub}{A string to remove off the data.frames if the names contain lots of unwanted duplication}
}
\description{
Takes text files generated from Open Clinica and does some initial formatting saving them in an R format.
}
\details{
\code{import.openClinica} creates a data.frame from a comma delimited file. Any column with the word date and any column named in the date variable in it is formatted to a \code{\link{POSIXct}} date

\code{}

(path,trialStub,dbName,newFolder,delStub="")
}
\section{Functions}{
\itemize{
\item \code{importOpenClinica_batch}: Runs importOpenClinica on a folder and saves to an .Rdata file.
}}

