% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_modified.R
\name{legend_modified}
\alias{legend_modified}
\title{legend_modified}
\usage{
legend_modified(x, y = NULL, legend, fill = NULL, col = par("col"),
  border = "black", lty, lwd, pch, angle = 45, density = NULL,
  bty = "o", bg = par("bg"), box.lwd = par("lwd"), box.lty = par("lty"),
  box.col = par("fg"), pt.bg = NA, cex = 1, pt.cex = cex,
  pt.lwd = lwd, xjust = 0, yjust = 1, x.intersp = 0.75,
  y.intersp = 0.75, adj = c(0, 0.5), text.width = NULL,
  text.col = par("col"), text.font = NULL, merge = do.lines && has.pch,
  trace = FALSE, plot = TRUE, ncol = 1, horiz = FALSE, title = NULL,
  inset = 0, xpd, title.col = text.col, title.adj = 0.5, seg.len = 2,
  xbox = 1, ybox = 1, title.cex = cex)
}
\arguments{
\item{pt.cex}{the sizing for the text}

\item{xbox}{Scales up the x-direction of boxes (generated when colours are provided to fill)Default 1.}

\item{ybox}{Scales up the y-direction of boxes (generated when colours are provided to fill). If this is 2 then the boxes will touch in vertical mode. Default 1.}

\item{title.cex}{the sizing for the title if provided.}
}
\description{
This is a modified version of the base function \code{\link{legend}}. In particular it adds xbox, ybox, title.cex and returns functionality to pt.cex which is broken in the base version.
}
