% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenClinica_OC_posIXct.R
\name{OC_posIXct}
\alias{OC_posIXct}
\title{Open Clinica as.POSIXct function}
\usage{
OC_posIXct(data)
}
\arguments{
\item{data}{A date generated by Open Clinica}
}
\description{
This is a wrapper for \code{\link{as.POSIXct}} to make formating dates consistently and easily from Open clinica databases.
}
