\name{peter.utilities-package}
\alias{peter.utilities-package}
\alias{peter.utilities}
\docType{package}
\title{
Peter's homemade utility functions
}
\description{
A collection of utility functions which extend basic functionality of base functions in R
}
\details{
\tabular{ll}{
Package: \tab peter.utilities\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-01-11\cr
License: \tab None\cr
}
This package includes the following functions:\cr
\code{\link{usePackage}}: Will install missing packages if missing before loading them using library.\cr
\code{\link{plot.two}}: plots two variables on one set of axes, the scales and names are added to the left and right axes.
}
\author{
Peter Dutton

Maintainer: Peter Dutton <peter.dutton@ndorms.ox.ac.uk>

}
\references{
None
}

\keyword{
}
\seealso{

}
\examples{
}
