% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_queryList_referenceClass.R
\name{queryList}
\alias{queryList}
\title{queryList}
\usage{
queryList(q = NULL, queryRun = 1)
}
\arguments{
\item{q}{query data.frame with the columns id, identifier, queryRun, patid, CRF, repeatLine, firstQuery, STATScomments,comments, resolved, STATSresolved}

\item{queryRun}{A number denoting the query run. Useful for seeing what queries are still active.}
}
\description{
A generator function for the \code{\link{dataQueries-class}}. Takes an optional dataframe and queryRun number and creates a reference object. If the data.frame is not provided then an empty one is generated.
}

