% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility; merge v0.1.R
\name{ut.merge}
\alias{ut.merge}
\title{Merge databases}
\usage{
ut.merge(merge.param, db1, db2, entries1, entries2)
}
\arguments{
\item{merge.param}{A vector of column names to use to merge data}

\item{db1, db2}{The data.frames to merge}

\item{entries1, entries2}{1 or "m" to tell it whether to do 1:1, 1:m or m:1 matching between databases}
}
\description{
Provides a function simular to merge in Stata to merge data from two data.frames.
}

